+++ rally_openstack/task/contexts/quotas/quotas.py
@@ -71,7 +71,7 @@ class Quotas(context.OpenStackContext):
                     # NOTE(andreykurilin): in case of existing users it is
                     #   required to restore original quotas instead of reset
                     #   to default ones.
-                    if "existing_users" in self.context:
+                    if "existing_users" in self.context["config"]:
                         self.original_quotas.append(
                             (service, tenant_id,
                              self.manager[service].get(tenant_id)))